Time: 0 ms (100%), Space: 19.6 MB (71.92%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.cpp
@@ -26,7 +26,7 @@ public:
             maxw = max(maxw, r-l+1);
             int n = q.size();
             while(n--){
-              int ind =  q.front().second;
+              long long ind =  q.front().second;
               TreeNode*curr = q.front().first;
        q.pop();
                    if(curr->left)
